requirements.txt
streamlit
moviepy
opencv-python-headless
pillow
pilmoji
numpy